readme

+++ Objects/ZBuffer.h
@@ -7,7 +7,7 @@
 class ZBuffer : public std::vector<std::vector<double>>{
 public:
     ZBuffer(const int &width, const int &heigth);
-    static double calculateVal(const int &zA, const int &zB, const int &a, const int &i);
+    static double calculateVal(const double &zA, const double &zB, const int &a, const int &i);
     bool apply(const int &x, const int &y, const double &zA, const double &zB, const int &a, int &i);
 };
